% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file_structure.R
\name{find_file_structure}
\alias{find_file_structure}
\title{Finds the file Structure of website}
\usage{
find_file_structure(years = NULL)
}
\arguments{
\item{years}{numeric vector. List of years in which to obtain file structure for. Default = NULL (find structure for all data)}
}
\value{
list of folders with filenames as contents
}
\description{
Find all subfolders containing data and lists the names of the files within each subfolder
}
\examples{
\dontrun{
To find the structure for years 2000, ..., 2010
fileStr <- find_file_structure(years = 2000:2010)

To find the structure for all data present
fileStr <- find_file_structure()
}

}
