% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_years.R
\name{find_years}
\alias{find_years}
\title{Finds the years in which data are available}
\usage{
find_years()
}
\value{
vector of years
}
\description{
Scans the oisst website for a list of years for which there are data
}
