% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oisst_data.R
\name{get_oisst_data}
\alias{get_oisst_data}
\title{Get oisst data}
\usage{
get_oisst_data(years = 2000, outputStructure = NULL, outputDir)
}
\arguments{
\item{years}{numeric vector. Years of data to download. Default = 2000. options: NULL (All data is downloaded)}

\item{outputStructure}{character. Determines how downloaded data will be structured. Default = NULL (All files downloaded to a single directory), options: "Year" (Data downloaded into Yearly folders),"Month" (Downloaded into Monthly Folders)}

\item{outputDir}{string. Full path to local machine where output files will be downloaded}
}
\value{
Nothing. Files downloaded
}
\description{
Pulls data from oisst website and downloads to local machine
}
\examples{
\dontrun{

Download years 2000, ..., 2010 into yearly folder within the "output" folder in current wd
get_oisst_data(years=2000:2010,outputStructure="year",outputDir=here::here("output"))

Download years 2000, ..., 2005 into a single folder within the "output" folder in current wd
get_oisst_data(years=2000:2005,outputStructure=NULL,outputDir=here::here("output"))

Download all data into monthly folder within the "output" folder in current wd
get_oisst_data(years=NULL,outputStructure="month",outputDir=here::here("output"))
}

}
